refactor: ExHandler 클래스 domain->ui 로 이동
