+++ src/main/java/nextstep/subway/applicaion/LineService.java
@@ -5,8 +5,8 @@ import nextstep.subway.applicaion.dto.LineRequest;
 import nextstep.subway.applicaion.dto.LineResponse;
 import nextstep.subway.applicaion.dto.LineUpdateRequest;
 import nextstep.subway.domain.*;
-import nextstep.subway.domain.exception.exception.LineNotFoundException;
-import nextstep.subway.domain.exception.exception.StationNotFoundException;
+import nextstep.subway.domain.exception.LineNotFoundException;
+import nextstep.subway.domain.exception.StationNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
