+++ src/main/java/nextstep/subway/domain/exception/LineNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain.exception.exception;
+package nextstep.subway.domain.exception;
 
 public class LineNotFoundException extends EntityNotFoundException {
